backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pdfplumber==0.11.0
python-docx==1.1.2
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
chromadb==0.5.15
sentence-transformers==3.1.1
google-generativeai==0.7.2
tiktoken==0.7.0
werkzeug==3.0.4
